+++ app/bot.py
@@ -1,12 +1,10 @@
 from urllib.parse import unquote
 
-from flask import Blueprint, abort, current_app, request
 from slack import WebClient
 
 from app.groups_read import get_groups_list
 from app.groups_write import invite_user_to_group, request_to_join_group
 
-route_blueprint = Blueprint("route_blueprint", __name__)
 
 PRIVATE_MESSAGE_NUDGE = "Please direct message me to get the list or join a channel. :slightly_smiling_face:"
 
@@ -24,7 +22,6 @@ def _is_private_message(form_data):
     return form_data["channel_name"] == "directmessage"
 
 
-@route_blueprint.route("/list", methods=["POST"])
 def list_groups():
     form_data = request.form
     if not _is_request_valid(
@@ -40,7 +37,6 @@ def list_groups():
     return get_groups_list(slack_web_client())
 
 
-@route_blueprint.route("/join", methods=["POST"])
 def join_channel():
     form_data = request.form
     if not _is_request_valid(
@@ -58,7 +54,6 @@ def join_channel():
     )
 
 
-@route_blueprint.route("/confirm_invite", methods=["GET", "POST"])
 def confirm_invite():
     auth_code = request.args["code"]
     client = slack_web_client("")
